admxgen: move in cmd/
Move admxgen command in cmd, and thus, rename admxgen module as an
internal content one.
This will use templates and everything local, and only expose 2 APIs
handling files and system.
